requirements.txt
Flask
Flask-SQLAlchemy
pandas
numpy
scikit-learn
joblib
mysql-connector-python
